moved symbolic probability expressions back to stats module
