+++ sympy/stats/symbolic_probability.py
